% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/em_function.R
\name{EM_PMM_algo}
\alias{EM_PMM_algo}
\title{Expectation Maximization Poisson Mixtures}
\usage{
EM_PMM_algo(x, K, iter = 1000)
}
\arguments{
\item{x}{this is the data}

\item{K}{this is the number}

\item{iter}{number of iterations defaults to 1000}
}
\description{
This function allows you to run an ML - EM algorithm for learning Poisson mixtures
}
\examples{
#generate some Poisson mixture data
data <- c(rpois(100,2), rpois(100,10))

#run the ML-EM algorithm to learn the mixture model
t1 <- EM_PMM_algo(data,2,iter = 100)

#plot the EM objective function, should be monotonically increasing and monitors convergence
plot(t1$sum_of_rows_ln_sum,
     type = 'l',
     main = "convergence of EM objective function",
     bty = 'n')

#check the cluster assignments (using original data this is not a prediction, which can also be done)
cluster_indicator <- c()
for(i in 1:length(data)){
  cluster_indicator <- c(cluster_indicator,which.max(t1$norm_phi[i,]))
}
plot(data,cluster_indicator,
     pch = 16,
     bty = 'n')
}
\keyword{em}
