% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vi_function.R
\name{VI_PMM_algo}
\alias{VI_PMM_algo}
\title{Variational Inference Poisson Mixture}
\usage{
VI_PMM_algo(x, K, iter = 1000, alpha = 0.1, a = 4.5, b = 0.25)
}
\arguments{
\item{x}{this is the data}

\item{K}{this is the number}

\item{iter}{number of iterations defaults to 1000}

\item{alpha}{hyperparameter of probability distribution of Poisson models defaults to 0.1}

\item{a}{Gamma prior parameter defaults to 0.1}

\item{b}{Gamma prior parameter defaults to 0.25}
}
\description{
This function allows you to run a VI algorithm for learning Poisson mixtures
}
\examples{
#generate some Poisson mixture data
data <- c(rpois(100,2), rpois(100,10))

#run the VI algorithm to learn the mixture model
t2 <- VI_PMM_algo(data,2,iter = 1000)

#plot the VI objective function, should be monotonically increasing and monitors convergence
plot(t2$L_vi,
     type = 'l',
     main = "convergence of VI objective function",
     bty = 'n')
     
#check the cluster assignments (using original data this is not a prediction, which can also be done)
cluster_indicator <- c()
for(i in 1:length(data)){
  cluster_indicator <- c(cluster_indicator,which.max(t2$norm_phi[i,]))
}
plot(data,cluster_indicator,
     pch = 16,
     bty = 'n')
}
\keyword{vi}
